image_resizer/requirements.txt
pika==1.1.0
Pillow==6.1.0
